Struct jibi::Interpreter

pub struct Interpreter { /* fields omitted */ }

Implementations

impl Interpreter

jibi Language Interpreter

Each Interpreter has its own state and global environment, and has methods to construct (interned) jibi values and procedures, define global variables, call procedures, evaluate expressions, and execute code.

pub fn new() -> Self

Create a bare jibi Interpreter, with no builtins or prelude; call Interpreter::default instead if you want those.

pub fn eval_tokens(
&mut self,
tokens: Box<dyn TokenIter>
) -> Result<Option<JValRef>, JException>

Evaluate a stream of tokens and return the value of the last expression, (or None if the program contains no expressions).

pub fn eval_str(
&mut self,
name: &str,
program: &str
) -> Result<Option<JValRef>, JException>

Evaluate a jibi script, and return the value of its last expression (or None if the program contains no expressions).

name: Name used to report errors in the program (e.g. filename, “stdin”).

program: One or more jibi expressions.

pub fn eval_file<P: AsRef<Path>>(
&mut self,
path: P
) -> Result<Option<JValRef>, JException>

Evaluate a jibi script file, and return the value of the last expression or None if the file contains no expressions).

path: Path to script file.

pub fn eval(&mut self, expr: JValRef) -> JResult

Evaluate a jibi expression. (All values are expressions, primitive types evaluates to themselves.)

pub fn call(&mut self, name: &str, args: Vec<JValRef>) -> JResult

Call a named jibi procedure.

name: Name of procedure (looked up in globals).

args: Vector of argument values.

pub fn print_exc((pos, err, tb): JException)

Print exception and traceback.

pub fn def(&mut self, name: &str, val: JValRef)

Create a global binding (variable definition).

name: Name to bind. It is possible to bind names which are not valid jibi symbols, they will not be accessible in jibi code.

val: A jibi value.

pub fn lookup(&self, name: &str) -> Option<JValRef>

Get value of binding in global environment.

name: Name to lookup. It is possible to bind names which are not valid jibi symbols, they will not be accessible in jibi code.

pub fn nil(&mut self) -> JValRef

Construct a jibi nil (always interned).

pub fn bool(&mut self, b: bool) -> JValRef

Construct a jibi bool (always interned).

pub fn int(&mut self, n: i128) -> JValRef

Construct a jibi integer (may be interned).

pub fn float(&mut self, v: f64) -> JValRef

Construct a jibi float.

pub fn symbol(&mut self, s: String) -> JValRef

Construct a jibi symbol (always interned).

pub fn string(&mut self, s: String) -> JValRef

Construct a jibi string (may be interned).

pub fn quote(&mut self, v: JValRef) -> JValRef

Construct a jibi quote.

pub fn list(&mut self, v: Vec<JValRef>) -> JValRef

Construct a jibi list (linked list made from pair and terminated by nil).

pub fn pair(&mut self, left: JValRef, right: JValRef) -> JValRef

Construct a jibi pair (cons cell).

pub fn error(&mut self, kind: JErrorKind, reason: &str) -> JValRef

Construct a jibi error value.

Note that JError can be found both in Ok(JValRef), as first class values that can be passed around in jibi, and in Err(JError) when it is raise’d by jibi code, or due to parsing or evaluation errors.

pub fn lambda(&mut self, params: Vec<String>, body: Vec<JValRef>) -> JResult

Construct a jibi lambda.

pub fn procmacro(&mut self, params: Vec<String>, body: Vec<JValRef>) -> JResult

Construct a jibi procmacro.

pub fn builtin<F>(&mut self, name: String, f: F) -> JValRef where
F: 'static + Fn(JValRef, JEnvRef, &mut JState) -> JResult,

Define a jibi builtin procedure.

pub fn specialform<F>(&mut self, name: String, f: F) -> JValRef where
F: 'static + Fn(JValRef, JEnvRef, &mut JState) -> JResult,

Define a jibi builtin special form.